+++ packages/reports/src/reports/blocks/types/renderers/Generator.jsx
