+++ packages/reports/src/reports/blocks/types/renderers/Stat.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import {Text} from "@mantine/core";
 
-import sanitizeBlockContent from "../../../utils/blocks/sanitizeBlockContent";
+import sanitizeBlockContent from "../../../../utils/blocks/sanitizeBlockContent";
 
 /**
  * "stat" block renderer
